requirements.txt
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
clarifai==2.2.3
Click==7.0
configparser==3.8.1
cryptography==2.9.2
decorator==4.4.2
Flask==1.1.1
future==0.18.2
googleapis-common-protos==1.51.0
grpcio==1.28.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==2.6.0
MarkupSafe==1.1.1
nltk==3.4.5
polyglot==16.7.4
protobuf==3.11.3
pycparser==2.20
PySocks==1.7.1
python-telegram-bot==12.5.1
requests==2.23.0
six==1.14.0
tornado==6.0.4
urllib3==1.25.9
Werkzeug==1.0.0
